[BLEEDING/INSTABLE] Roughly adjust to 1.6.1 Bukkit API changes.

+++ NCPCore/src/main/java/fr/neatmonster/nocheatplus/checks/fight/FightData.java
@@ -113,7 +113,7 @@ public class FightData extends ACheckData {
     
     // New god mode check [in progress].
 	public int					   godModeHealthDecreaseTick 	= 0;
-	public int                     godModeHealth       			= 0;
+	public double                  godModeHealth       			= 0.0;
     public int                     lastDamageTick 				= 0;
     public int                     lastNoDamageTicks 			= 0;
     /** Accumulator. */
